celery: change celery startup to limit memory usage
Currently celery in forked from webhookd and this after webhookd have
been fully setuped, this allocate ton of useless resources in celery
workers.

This change loads celery earlier and only load the tasks module and
nothing else.

Configuration options have also changed to the new celery 4.0 format.
Some have been removed as they are now the defaults.
