+++ wazo_webhookd/controller.py
@@ -2,26 +2,29 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
-import os
 import signal
 
 from functools import partial
-from multiprocessing import Process
 from threading import Thread
 from xivo import plugin_helpers
 from xivo.consul_helpers import ServiceCatalogRegistration
 from wazo_auth_client import Client as AuthClient
 
 from .bus import CoreBusConsumer
-from .celery import CoreCeleryWorker
-from .celery import app as celery_app
 from .rest_api import api, CoreRestApi
+from wazo_webhookd import celery
 
 logger = logging.getLogger(__name__)
 
 
 class Controller:
     def __init__(self, config):
+        # NOTE(sileht): Celery must be spawned before anything else, to ensure
+        # we don't fork the process after some database/rabbitmq connection
+        # have been established
+        celery.configure(config)
+        self._celery_process = celery.spawn_workers(config)
+
         self._service_discovery_args = [
             'wazo-webhookd',
             config.get('uuid'),
@@ -30,12 +33,9 @@ class Controller:
             config['bus'],
             lambda: True,
         ]
+
         self._auth_client = AuthClient(**config['auth'])
         self._bus_consumer = CoreBusConsumer(config)
-        self._celery_worker = CoreCeleryWorker(config)
-        self._celery_process = Process(
-            target=self._celery_worker.run, name='celery_process'
-        )
         self.rest_api = CoreRestApi(config)
         self._service_manager = plugin_helpers.load(
             namespace='wazo_webhookd.services',
@@ -43,7 +43,6 @@ class Controller:
             dependencies={
                 'api': api,
                 'bus_consumer': self._bus_consumer,
-                'celery': celery_app,
                 'config': config,
                 'auth_client': self._auth_client,
             },
@@ -62,7 +61,6 @@ class Controller:
     def run(self):
         logger.info('wazo-webhookd starting...')
         signal.signal(signal.SIGTERM, partial(_sigterm_handler, self))
-        self._celery_process.start()
         bus_consumer_thread = Thread(
             target=self._bus_consumer.run, name='bus_consumer_thread'
         )
@@ -73,7 +71,7 @@ class Controller:
         finally:
             logger.info('wazo-webhookd stopping...')
             self._bus_consumer.should_stop = True
-            os.kill(self._celery_process.pid, signal.SIGTERM)
+            self._celery_process.terminate()
 
             logger.debug('waiting for remaining threads/subprocesses...')
             bus_consumer_thread.join()
